+++ tibiapy/models/character.py
@@ -228,9 +228,6 @@ class Killer(BaseModel):
 
 class OnlineCharacter(BaseCharacter):
     """An online character in the world's page."""
-
-    world: Optional[str]
-    """The name of the world."""
     vocation: Vocation
     """The vocation of the character."""
     level: int